+++ pdr_backend/data_eng/test/test_model_factory.py
@@ -3,8 +3,8 @@ import warnings
 from enforce_typing import enforce_types
 import numpy as np
 
-from pdr_backend.model_eng.model_factory import ModelFactory
-from pdr_backend.model_eng.model_ss import APPROACHES, ModelSS
+from pdr_backend.data_eng.model_factory import ModelFactory
+from pdr_backend.data_eng.model_ss import APPROACHES, ModelSS
 
 
 @enforce_types
