+++ pdr_backend/predictoor/approach3/predictoor_agent3.py
@@ -4,7 +4,7 @@ from enforce_typing import enforce_types
 
 from pdr_backend.data_eng.data_factory import DataFactory
 from pdr_backend.data_eng.data_pp import DataPP
-from pdr_backend.model_eng.model_factory import ModelFactory
+from pdr_backend.data_eng.model_factory import ModelFactory
 
 from pdr_backend.predictoor.base_predictoor_agent import BasePredictoorAgent
 from pdr_backend.predictoor.approach3.predictoor_config3 import PredictoorConfig3
