+++ pdr_backend/simulation/trade_engine.py
@@ -10,8 +10,8 @@ from statsmodels.stats.proportion import proportion_confint
 from pdr_backend.data_eng.data_factory import DataFactory
 from pdr_backend.data_eng.data_pp import DataPP
 from pdr_backend.data_eng.data_ss import DataSS
-from pdr_backend.model_eng.model_factory import ModelFactory
-from pdr_backend.model_eng.model_ss import ModelSS
+from pdr_backend.data_eng.model_factory import ModelFactory
+from pdr_backend.data_eng.model_ss import ModelSS
 from pdr_backend.simulation.sim_ss import SimSS
 from pdr_backend.simulation.trade_ss import TradeSS
 from pdr_backend.simulation.trade_pp import TradePP
